pyproject.toml
[tool.poetry]
name = "german-law-bot"
version = "0.1.0"
description = ""
authors = ["Felix Ocker <felix.ocker@googlemail.com>"]
readme = "README.md"
packages = [{include = "german_law_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
chromadb = "^0.4.6"
black = "^23.7.0"
ruff = "^0.0.285"
gradio = "^3.40.1"
pyyaml = "^6.0.1"
openai = "^1.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
